requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
greenlet==3.3.2
pymysql==1.1.0
aiomysql==0.2.0
cryptography==42.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==7.0.0
httpx==0.26.0
alembic==1.13.1
email-validator==2.1.0
